xsctapp: use common repo/branch wiring
This commit moves repository, branch, and commit wiring into a common
mechanism in xsctapp.bbclass.  This way, individual applications do not
need to maintain custom ways to get to a common embeddedsw repository.

In addition, we also rewire the deploy content to properly include
versioning information and symlinks to allow multiple pmu/fsbl/fsboot
binaries in the deploy area.

In future, this also allows for a more elegant shared source mechanism
that can allow a single repo to be reused across recipes just as with
the gcc/kernel recipes.

Signed-off-by: Cyril Chemparathy <cyril.chemparathy@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>