xilinx-bootbin.bbclass: fix task dependencies when using wic
Image recipes create rootfs' which aren't generally included in a
boot.bin. So there is no reason for the bif generation to wait for
do_image_complete to be done. The bif creation is already bound to it's
own task dependencies using the depends var flags populated by the
get_bootbin_depends return value.

Inserting the boot.bin generation between do_image_complete and do_build
makes it impossible for do_image_wic to find the boot.bin in the
DEPLOY_DIR. This is because do_image_wic (and _wicenv) is performed
before image do_build.

Signed-off-by: Jean-Francois Dagenais <jeff.dagenais@gmail.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>