externalxsctsrc.bbclass: Provide src and/or build option to check file hashes
When using files from an external source, bitbake should spend time to
hash all the files to know if any of the files have been changed. This
results in builds taking more time. Incase of petalinux-tools, since the
external source is coming from the tool itself, we can skip hashing the
source files  when building and only hash the files in the build
directory.

This patch gives the option of specifiying whether we need to add src
and/or build files to the hash check list. By setting
EXTERNALXSCTSRCHASH to src and/or build we can control what files gets
added to the hash check list.

By default it will hash the source files when EXTERNALXSCTSRCHASH is
empty

Signed-off-by: Bhargava Sreekantappa Gayathri <bhargava.sreekantappa-gayathri@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>