xilinx-mcs_1.0.bb: Add mcs generation for microblaze
25 years ago, eeproms were programmed with prom programmers, that had
Intel format, Motorola format, and various other formats.  The eeproms
made by Intel and Motorola are long gone, and so are the programmers,
but the file formats remain, and became defacto standards. The .mcs,
.rbt, and others are ASCII representations (ASCI 1 = 31H, and ASCII
0=30H) with headers, comments, etc. The .bin, .bit are binary files:
all single bits, 1's and zeros. All formats have evolved over the years,
and are still used.  New companies make programmers, and all the old
formats get supported.

So a .bin is used to program the part using an
Impact programming cable (JTAG or parallel), and a .rbt is the file
readback when doing a verify Impact (also provide a .bin readback file
which is not equal to the .bin file used to program the part!).

The .mcs might get used to program a prom for the final board, to
configure the part upon power on.

This patch adds a recipe to generate mcs file for microblaze using
Vivado.

Signed-off-by: Bhargava Sreekantappa Gayathri <bhargava.sreekantappa-gayathri@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>