xsct-tc.bbclass: Gracefully exit if XILINX_SDK_TOOLCHAIN is not set
XILINX_SDK_TOOLCHAIN is set in local.conf to point to Xilinx SDK
installation directory. Currently, if this variable is not defined
anywhere, the build fails and the error message looks like this:

line 27, in xsct_get_version(d=<bb.data_smart.DataSmart object at
0x7f845a368a90>):
         else:
    >        last_line = stdout.splitlines()[0].split()[-2]
             return last_line[1:7]
IndexError: list index out of range

The reason for this is, we are calling xsct_run in a try and except
block. The idea being, "return bb.process.run(cmd, cwd=topdir,
stderr=subprocess.PIPE) " will raise an exception which can be
intercepted in the except block. But, if XILINX_SDK_TOOLCHAIN is not
set, then we are returning UNKNOWN to both stdin and stderr and not raising
any exception. This patch will fix this issue by making sure the stdout
is not UNKNOWN before parsing it. do_xsct_setup will exit gracefully
with the right message in this case.

Signed-off-by: Bhargava Sreekantappa Gayathri <bhargava.sreekantappa-gayathri@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>