layer.conf: inherit xsct-tc
xsct-tc.bbclass contains eventhandler to check for xilinx tools version.
This eventhandler is configured to trigger for BuildStarted event by
setting eventmask. For do_xsct_setup eventhandler to see this event,
xsct-tc.bbclass needs to be inherited in a config file and not in a
recipe.

Previously xsct-tc was getting inherited in xilinx-bootbin recipe and
hence it was not seeing the BuildStarted event and the eventhandler was
not working. This patch changes this behaviour by inheriting from the
layer.conf so that when this layer is included, the eventhandler is
enabled by default.

Signed-off-by: Bhargava Sreekantappa Gayathri <bhargava.sreekantappa-gayathri@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>