fsbl_git.bb: Remove DXPS_BOARD flags for ZCU102 and ZCU106
For production versions of ZCU106 and ZCU102, the board name is set by
passing a DXPS_BOARD flag to the hardware flow scripts through the fsbl
recipe.

The designs from PLNX and hardware flow scripts incorporate presets to
set the BOARD name and no longer require the flags in fsbl code. This
patch removes the flags.

Signed-off-by: Sai Hari Chandana Kalluri <chandana.kalluri@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>