bitstream-microblaze_1.0.bb: Add download.bit for microblaze
If the elf file you want to run on the FPGA is small enough to run in the BRAM,
then you can take the bit file and include the elf file in it and produce a
download.bit. When this file is programmed to the FPGA, the elf binary will
automatically run after the FPGA is programmed.

A common use case is adding a bootloader to the the bitfile so that the
bootloader will start after programming the FPGA and runs a much bigger elf
from external memory(like u-boot).

Download.bit can be generated using updatemem tool.

This patch will add a recipe to generate download.bit for microblaze.

Signed-off-by: Bhargava Sreekantappa Gayathri <bhargava.sreekantappa-gayathri@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>