xsctyaml.bbclass: Adding flag to support device tree overlay
Adding a flag called YAML_ENABLE_DT_OVERLAY that will add 'dt_overlay' to
YAML_BSP_CONFIG if set. With this flag enabled, pl.dtsi will not be
included in the base dts and instead will have a '/plugin/' tag which
will enable logic to have it compiled separately as a dtbo.

Signed-off-by: Jaewon Lee <jaewon.lee@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>