bitstream-microblaze_1.0.bb, xilinx-mcs_1.0.bb : Fix do_deploy basehash mismatch
This patch fixes the below issue :

ERROR: When reparsing
/yocto1/sources/core/../meta-xilinx-tools/recipes-bsp/bitstream/bitstream-microblaze_1.0.bb.do_deploy,
the basehash value changed from ... to .... The metadata is not
deterministic and this needs to be fixed.

bitstream-microblaze can be used seperately without generating boot.mcs,
so add deploy variables in original recipe. Also, use the same variables
in do_deploy task.

Signed-off-by: Sreeja Vadakattu <svadakat@xilinx.com>
Signed-off-by: Sai Hari Chandana Kalluri <chandana.kalluri@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>