xilinx-bootbin: Explicitly set the BIF_PARTITION_ATTR
BIF_PARTITION_ATTR requires components of BOOT.bin in a specific order
i.e fsbl, bitstream, pmufw, atf, u-boot. The fsbl loads the bitstream
and pmufw, however in the current recipe the bistream is appended to
the end of the BIF_PARTITION_ATTR order. During SD boot, fsbl waits to
load the bitstream and consequently fails. This patch sets the
BIF_PARTITION_ATTR to the correct order.

Signed-off-by: Sai Hari Chandana Kalluri <chandana.kalluri@xilinx.com>
Signed-off-by: Jaewon Lee <jaewon.lee@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>