xsct-tarball.bbclass: adding flag VALIDATE_XSCT_CHECKSUM
Adding a validate checksum flag to toggle a fatal error if checksum is
not the same as the one defined in the class. Default behavior is to
compare the checksums. Also changing chksum_tar to chksum_tar_actual
to go along with this change as chksum_tar is no longer guaranteed to be
the actual checksum of the tarball.

Signed-off-by: Jaewon Lee <jaewon.lee@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>