fpga-manager-util_1.0.bb: Add HDF_LIST to optimize compile and install
Make use of HDF_LIST, which is a list of all extra hdf names, to be able
to loop only once in both compile and install functions. Instead of
having two loops for dtsi/dtbo and bit/bin files, there will be one
iteration through the hdf names to find the appropriate files.

When adding HDF_LIST, there was a very similar variable called
extrapackages, where the only difference was a prepended '${PN}-'
Defining extrapackages at the end, with use of format() function to
simplify code.

Signed-off-by: Jaewon Lee <jaewon.lee@xilinx.com>
Signed-off-by: Alejandro Enedino Hernandez Samaniego <alejandr@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>