fpga-manager-util_1.0.bb: Add extra hdfs to SRC_URI
Adding all hdfs in EXTRA_HDF path to src_uri in an anonymous python
function to support auto rebuilding if an hdf in that directory is
added/removed.

Cleaning necessary directories before the fetch and configure tasks to
ensure removed hdfs are not packaged. (This will mean any change of hdfs
in EXTRA_HDF will trigger a clean build) All items in src_uri are copied
to $WORKDIR, so in this case all hdfs will be stored in
${WORKDIR}${EXTRA_HDF}. Setting XSCTH_HDF to ${WORKDIR}${EXTRA_HDF} to
use hdfs from within the WORKDIR.

Signed-off-by: Jaewon Lee <jaewon.lee@xilinx.com>
Signed-off-by: Alejandro Enedino Hernandez Samaniego <alejandr@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>