device-tree.bbappend: Add device-tree nodes for MIPI support on Ultra96
To enable the MIPI camera support on Ultra96, the OV5640 and video
pipeline nodes must be included in the device tree. This patch includes
mipi_for_ultra96.dtsi containing the required nodes. At the end of
configure step, it copies the provided dtsi to build dir and appends the
file to system-top.dtsi

Signed-off-by: Sreeja Vadakattu <svadakat@xilinx.com>
Signed-off-by: Sai Hari Chandana Kalluri <chandana.kalluri@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>