xsct-tarball.bbclass: check ext_tarball is set before copying to dl dir
This fix is required to bypass error that will happen in the esdk
creation when including xsct-trim. Upon creation, ext_tarball will not
be set for the esdk so there will be a fatal error at the shutil.copy
line. Checking to make sure ext_tarball is set before copying to dl dir

Signed-off-by: Jaewon Lee <jaewon.lee@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>