xsct-tarball.bbclass: Fix extraction logic for when not validating chksum
Fixing extraction logic for when VALIDATE_XSCT_CHECKSUM is set to '0' In
a previous patch, we switched chksum_tar_actual to chksum_tar because in
the validating checksum path, those two are guaranteed to be the same.
In the case where VALIDATE_XSCT_CHECKSUM is '0' however they may be
different. Also setting chksum_tar_actual to the recipe defined chksum
if fetching the tarball as we cant do a md5sum on the file from a
webserver. This is safe however because if chksum_tar_recipe is the
wrong value, the fetcher will fail later with a md5sum error.

Also renamed chksum_tar to chksum_tar_recipe to avoid confusion

Signed-off-by: Jaewon Lee <jaewon.lee@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>