Added link for providing a boot.bin in deploy
If this recipe is being used, it's the preferred provider of
virtual/boot-bin and should provide BOOT.bin or boot.bin.
If used (correctly) with the meta-xilinx layer, the
u-boot-zynq-uenv recipe from that layer will errantly
pick BOOT.bin and treat it like an FPGA image.  So, this
patch produces a boot.bin in the deploy directory, fixing
one problem while avoiding the other.

Signed-off-by: Thomas Goodwin <btgoodwin@geontech.com>