fpga-manager-util_1.0.bb: cp bin file to expected dir for zynq platform
Zynq platform needs to run bootgen with the -process_bitstream flag
which ignores the -o input to specify where to put the output and
instead just creates the bin file in the same directory as the
bitstream. copying the bin file to the current working directory when
building for zynq so the bin file exists where the rest of the recipe
expects it to be.

Signed-off-by: Jaewon Lee <jaewon.lee@xilinx.com>