Adding FPGA_MNGR_RECONFIG_ENABLE to control enabling fpga manager
Previously we were checking that 'fpga-manager' was enabled in
IMAGE_FEATURES when doing fpga-manager specific steps (enabling overlays
in dtg, taking out bitstream from boot.bin, etc). Introducing a new flag
called FPGA_MNGR_RECONFIG_ENABLE to be able to decouple from the
meta-petalinux layer that defines the IMAGE_FEATURE.

To be able to use fpga-manager with only meta-xilinx and
meta-xilinx-tools FPGA_MNGR_RECONFIG flag has to be set and
fpga-manager-script and fpga-manager-utils recipes added to the image
manually.

Signed-off-by: Jaewon Lee <jaewon.lee@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>