pmu-firmware.bb:Provide power button support for both V1 and V2 Ultra96-zynqmp boards
Ultra96 has a power button controller LTC2950 which generates an
active low interrupt on MIO26 whenever the power button is pressed.
Since PMU GPI (connected to MIO26) only supports positive edge triggered
interrupts, it is not possible to reliably capture the power button
press event. So, we choose to capture the button press event by polling
for the GPI pin state at regular intervals. When a LOW state is detected
on the pin, system shutdown process is initiated. As per the datasheet,
LTC2950 generates an active low pulse of minimum 26ms duration. So, to
ensure we capture the event reliably, a scheduler task that runs every
10ms is added to check for the pin state and initiate shutdown. This
patch adds Ultra96 module which implements this functionality and is
enabled by defining ENABLE_MOD_ULTRA96 flag. It depends on scheduler
feature which needs to be enabled (ENABLE_SCHEDULER).

If FSBL is not yet running, it means that MIO is not configured and the
GPI1 MIO bits in IOModule always read zero. This could cause false
triggers. Also PM module may not have the right IPI info. So skip pin
monitoring until config object is loaded.

For Ultra96 V2 board, the power button controller generates an active
high pulse. This polarity change between V1 and V2 boards is handled
based on ULTRA96_VERSION macro which needs to be set to 1 or 2, based on
the board version

By default ULTRA96_VERSION ?= 1 i.e enabled for V1 boards. To enable
power button support for V2 boards the user must set ULTRA96_VERSION=2
within conf/local.conf file.

Signed-off-by: Jyotheeswar Reddy <jyothee@xilinx.com>
Signed-off-by: Sai Hari Chandana Kalluri <chandana.kalluri@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>