xsct-tarball: Exclusively deal with XSCT only if it is necessary
The process of checking if the xsct-trim tarball is up to date
and/or fetching a new one is not only very slow and tedious, but
also unnecessary in many cases.

At this moment, if a user wishes to perform almost any task using
bitbake, the check for XSCT gets executed, whether or not XSCT was
actually required.

Since only a handful of packages require xsct to be built, we only
need to execute this task if a required package is meant to be built,
and even then, not all tasks from that package may need it.

The need for xsct comes from the class xsct-base, which overrides
the do_configure() task, hence, only tasks that depend on do_configure
(and do_configure itself) belonging to packages that inherit xsct-base
actually require XSCT.

Exclusively execute the check for XSTC only when the do_configure task
for one of the specified packages is to be populated on the runqueue,
avoid inflicting pain on users that do not require it.

The only downside of this, is that we need to keep a list of packages
that do require XSCT to be built, this is fine up to a certain extent,
although, if this list grows too much, it may be worth to use a method
to check for packages that inherit the xsct-base class instead.

Signed-off-by: Alejandro Enedino Hernandez Samaniego <alejandr@xilinx.com>