device-tree.bbappend: Modify system controller BOARD configuration
Use BOARD zynqmp-a2197-p-reva as YAML_DT_BOARD_FLAGS for system
controller. The existing configuration does not contain all the required
sd card nodes in the device tree and as a result during boot time the
kernel is unable to load the rootfs from the root partition.
zynqmp-a2197-p-reva provides all the required nodes ensuring the sdboot
is sucessful.

Signed-off-by: Sai Hari Chandana Kalluri <chandana.kalluri@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>