fpga-manager-util_1.0.bb: Allow for custom pl nodes for extra hdfs
With this patch, users can add custom pl nodes for the extra hdfs being
supplied to fpga-manager-util by putting the dtsi file into the same
directory the HDFS are (EXTRA_HDFS), and naming the dtsi file as
${HDF_NAME}.dtsi. For example, if EXTRA_HDF has two hdfs called one.hdf
and two.hdf, users can add nodes into the dtbos generated by putting a
one.dtsi and two.dtsi into the EXTRA_HDF directory.

The extra dtsi files are copied over to the work directory of each HDF
after do_configure and a pl-final.dts is compiled, which includes both
the base pl.dtsi and the copied over dtsi, now called pl-custom.dtsi.
These custom pl.dtsi files are also added to SRC_URI in the packaging
anonymous python function to be able to rebuild on changes.

One limitation with this flow is, because processing for all hdfs are
done under this one recipe, there is one yaml file supplied to the xsct
call and therefore separate BSP/APP configs cannot be specified per
extra hdf.

Signed-off-by: Jaewon Lee <jaewon.lee@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>