Settings YAML_OVERLAY_CUSTOM_DTS using = instead of ?=
Setting YAML_OVERLAY_CUSTOM_DTS to pl-final.dts using = instead of ?=
as default flow will now build and check for a pl-final.dtbo. There is
no need for users to change this variable. We are now building a
pl-final.dtsi that includes both pl.dtsi and pl-custom.dtsi to provide a
way for users to inject custom nodes, but this doesn't necessarily mean
there has to be custom nodes

Signed-off-by: Jaewon Lee <jaewon.lee@xilinx.com>