device-tree.bb: Setting EXTERNALSRC_SYMLINKS to disable link creation
When using devtool, softlinks pointing to logs and workdir of the
bitbake workspace are populated in the devtool workspace for ease of
access. There is an hsi specific issue where it can't handle softlinks
in the work directory. Setting EXTERNALSRC_SYMLINKS to disable softlink
creation of directories like oe-workdir and oe-logs to bypass this
issue.

Signed-off-by: Jaewon Lee <jaewon.lee@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>