base-pdi_1.0.bb: Supporting designs with multiple pdis
Previously installation of the base-pdi was failing when the design had
multiple pdis. This patch uses the variable BASE_PDI_NAME to set the
base pdi name to allow the user to bypass this issue of not knowing what
the base pdi will be called in the design. If there is only one pdi,
that one will be installed as the base-pdi.

If the pdi file defined by BASE_PDI_NAME does not exist, there will be a
fatal error and it will list the possible choices. Users can set the
BASE_PDI_NAME variable in local.conf

Signed-off-by: Jaewon Lee <jaewon.lee@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>