xsct-tarball.bbclass: Improve hash checking speed
When checking the hash of the existing XSCT tarball to see
if it is still valid, due to the fact that its a very large
file (several GBs), it takes a long time to read the file
into memory by the hashlib python module, and then calculate
its md5sum, performing several experiments on different hosts
it ranges from 30-40 seconds to a maximum of about 320 seconds
averaging around 150, this is depending on how the file is
cached in memory and its affected by other processes that are
currently running on the system.

Profiling the code, we can see that the most amount of time is
spent reading the file and caching it into memory.

By reading the file in chunks, depending on block size multiplied
by a power of 2, we can see a huge improvement on calculating
this checksum, specially the second time this happens (after
bitbake has already been run), sometimes taking less than
10 seconds to go through this code.

Signed-off-by: Alejandro Enedino Hernandez Samaniego <alejandr@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>