xsct-tarball: Use a single xsct extracted toolchain on decoupling
Since we are now able to build a full system using the decoupling flow
several of the multiconfigs might require to use the XSCT toolchain for
the device-tree dependency still.

There is no point in having different copies of the same XSCT extracted
toolchain.
Allow the XSCT toolchain to be extracted on TOPDIR so all the multiconfigs
can use the same copy of it.

Signed-off-by: Alejandro Enedino Hernandez Samaniego <alejandr@xilinx.com>