**TEMPORARY** device-tree.bbappend: Override dtg generated pl.dtsi with required pl.dtsi
Override dtg generated pl.dtsi with copy of pl.dtsi that works with current
ultra96 xsa containing the mipi camera capture pipeline.  This fix will allow
device-tree compilation to be successful as all the required nodes in pl are
found correctly.

The correct fix would be to update the ultra96 xsa design and the
mipi-support-ultra96.dtsi file with correct pl nodes.

Signed-off-by: Sai Hari Chandana Kalluri <chandana.kalluri@xilinx.com>