fpgamanager_dtg.bbclass: Adding fpgamanager class for devicetree flow
Users will use this class when going through the xilinx device tree
generator flow. Inputs are an XSA, and optionally a custom dtsi
(to be included in the pl-final.dts) as well as an xclbin file, all
through SRC_URI

The bit file will be processed to .bit.bin and pl.dtsi will be extracted
from the extracted and compiled to dtbo along with the custom dtsi.
The above plus the optional xcl bin will be packaged and installed under
/lib/firmware/${PN}

Signed-off-by: Jaewon Lee <jaewon.lee@xilinx.com>