Removing IP instances and using IP names
The current xsct hsi APIs are using the peripheral defines(IP instances).
But for CIPS 3.0 and later hier designs there is a change in instance name property.
Due to this change there are failures in.
So to fix this using Canonical names(IP names) instead of peripheral defines.
In future if there is any change in instance name, there won't be any failures.

Signed-off-by: Varalaxmi Bingi <varalaxmi.bingi@xilinx.com>