openamp: Use processor instance name for openamp zynq platform
With recent changes on processor instance name to processor_ip name
causing the openamp run time issues, since this will runs on
instance 1 and app.tcl defaulting to instance 0.

This patch will add provision to specify processor_instance name using
XSCT_PROC variable. if this is defined, uses this for xsct to create app
instead of getting this from processor_ip so that it can support
cips3.0 also.

Signed-off-by: Raju Kumar Pothuraju <raju.kumar-pothuraju@xilinx.com>