embeddedsw: Adjust the pmu/plm/psm firmware to latest standalone
The standalone implementation has changes to allow standalone and XSCT
versions to work interchangably.  The changes also permit the XSCT
version to build as baremetal or Linux.  In the case of Linux we use
the XSCT embedded microblaze toolchain, but in the baremetal case
we build our own toolchain.

The build and formatting of things is controlled by the system
configuration.  However, inclusion of this layer should allow things
to work in most cases without any additional modifications.

Individual changes:
- xsct-tc: Allow this to work either xsct toolchain or baremetal
- xsctapp: Move source references to meta-xilinx-standalone and
            xlnx-embeddedsw.bbclass
- xsctapp: change 'make' to oe_runmake
- xsctbase: All B & S to be overwritten
- xsctbase: XSCTH_PROC uses an intermediate to allow recipes to easily
            override the default value(s)
- layer.conf: inherit xsct-tarball, not xsct-tc
- plm/psm/pmu: Split fw generation and packaging
- plm/psm/pmu: optimize skipping default dependencies
- external-hdf: optimize skipping default dependencies

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>