imgsel_git.bb: Adjust make call
in 2020.2.2_SOM there is command in compile task but in 2021.1 we got
oe_runmake which is not generating outputs.  so as workaround adding the
make command in do_compile task of 2021.1 to generate the outputs.

Signed-off-by: Swagath Gadde <swagath.gadde@xilinx.com>

Instead of replacing 'oe_runmake' with 'make', we match behavior by setting
the PARALLE_MAKE to "-j 1".

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>