xilinx-bootbin: Disabling the package generation steps triggers a warning
The following warning occurs:

    WARNING: u-boot-zynq-uenv-1.0-r0 do_package: Manifest manifest-x86_64_x86_64-nativesdk-xilinx-bootbin.packagedata not found in ... (variant '')?

Because u-boot-zynq-uenv has a dependency on xilinx-bootbin, it requires
access to the packagedata manifest file.  This is generated by the
do_packagedata step, but that step requires do_package to run first.

A prior change had already set 'PACKAGE = ""', so no package will be
generated, but at least the packagedata is now generated to avoid
the warning.

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>