device-tree: add variable EXTRA_OVERLAYS for overlaying system-top.dts at build time.
If a user is setting the variable EXTRA_OVERLAYS with paths to dtsi overlay
files, then each dtsi file will be applied on top of the system-top.dts file
in do_configure().

This can be used for a user of xilinx layers to easily add a device tree
overlay that can be applied on top of the system-top.dts file.

Signed-off-by: Ben Levinsky <ben.levinsky@xilinx.com>
Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>