xilinx-bootbin: Fix failure to find boot.scr file
In select configurations, the u-boot-zynq-scr:do_deploy was delayed until
after the xilinx-bootbin was run.  To avoid this issue, add an explicit
dependency on that recipe/task to the do_compile step.  Note, since this is
only needed for versal, we also need to subsequently limit the dependency
to only happen on versal.

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>