dfx-mgr: Update to latest 2022 version, adjust integration
Move the branch to master and SRCREV to the current 2022 version.

Add XRT as a stated dependency (zocl could bring this in).

Adjust the EXTRA_OECMAKE to pass in the recipe sysroot, so that we can find
the XRT headers.

The package also adds a new "libdfxgraph" library.  Add the library, the
associated header file, and split the existing dfx-mgr, libdfx-mgr, and
new libdfxgraph into individual packages.  This should allow only the
references components to be installed and used.

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>