fpgamanager_dtg_dfx.bbclass: New bbclass for DFX support
This class is to be used for partial XSAs when going through the
xilinx device tree generator flow and currently only supports
Versal. It requires an XSA file as an input. Optional inputs
include a .xclbin, a .dtsi, and an accel.json through SRC_URI.

The dfx scripts add support for partial XSAs and are used by the
fpgamanager_dtg_dfx class.

Signed-off-by: Jaewon Lee <jaewon.lee@xilinx.com>
Signed-off-by: Vishal Patel <vishal.patel@xilinx.com>
Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>