fpgamanager_dtg: Split SRC_URI with space
Splitting with file:// is not extracting the file name when git path is
in between (Ex: file://rp.xsa git://devicetree-xlnx.git file://base.tcl)
with above example the xsa path will be "rp.xsa
git://devicetree-xlnx.git" which is not proper.

To fix this splitting the SRC_URI with space and removes the prefix
'file://'.

Signed-off-by: Raju Kumar Pothuraju <raju.kumar-pothuraju@xilinx.com>
Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>