xsct-tarball.bbclass: Consider multiconfigs in check
We need to evaluate both regular and multiconfigs, but since the comparison
only affects the recipe.task, we need to split on ':' and take the last
part of the split.

The list comprehension was changed to a regular for loop as well for
performance and readability.  On no match this should take the same amount
of time, however on a match should stop after the first match.

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>