hdf-repository-generate-srcuri: Rework automation
The change to the directory structure requires a rework of the automation.

Iterate over the repository and break things up by directory.  Each file is
considered to be an XSA file.

Additionally, restructure how the replacement works to ensure that if this
set is partial, it can still all back to the default or other URIs.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>