layer.conf: Move to Yocto Project best practices
There is now an activation required for this layer.  Set:
     XILINX_WITH_ESW = 'xsct'

All of the recipe bbappends were modified to pay attention to that setting
as well.

The above modifications were based on meta-virtualization's approach.

There were two copies (the right and obsolete) versions of the
xsct_*_app.bbclass.  Only keep the version in the classes-recipe.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>