xsct-native: Move from class to recipe install
Remove the obsolete xsct-tarball approach.  Instead use a recipe to download and
install the xsct-trim components.  This contains a custom download and unpack
routine so that the prior behave can be mostly preserved.

Note, on a local path, the checksum is now automatically computed and no longer
verified.  This matches standard YP behavior.  A local file is copied to
the download directory.  This matches the current behavior and is slightly
different then the normal YP behavior.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>
(cherry picked from commit 60b029c565802aa759ecb8dea7e37ca69bb341b3)
Signed-off-by: Mark Hatle <mark.hatle@amd.com>