scripts/mb-realoc: reading payload address
In 2024.1 we started using u-boot.elf with is having relocation
and earlier we use u-boot without relocation.
In earlier binary the payload address used to come in .text
section and in the u-boot.elf it was in .data section.
So doing the changes.
Without this change fs-boot is unable to find the
u-boot load address in the header and mcs boot is failing.

Signed-off-by: Varalaxmi Bingi <varalaxmi.bingi@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>