xsct-tarball.inc: Ensure XILINX_SDK_TOOLCHAIN is valid at all times
One xsct-native is run (manually or through a dependency) an external script
may need to access the xsct-trim components.  Ensure that the path to the
components directory is defined by default.  However, when we are in a
recipe context, we should not be calling it from the components directory,
but instead from the recipe's recipe-sysroots-native directory.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>