xsct-native: Implement USE_XSCT_TARBALL workflow
The default behavior of using an XSCT tarball, either from  file or as a
download worked fine before.  However, disabling this was not implemented
but is a key workflow for PetaLinux.

Add support to disable the download and simply point to the location
specified by the user.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>
(cherry picked from commit c040ccd3398fbaf0a6b85180433f7934aa88ee44)
Signed-off-by: Mark Hatle <mark.hatle@amd.com>