Revert "meta-xilinx-tools: add ncurses-native to LD_LIBRARY_PATH for xsct"
This reverts commit d8b651d8d904f2913fc538325868471c81a13eef.

Adding ncurses-native to provide libtinfo.so.5 seems like a good idea,
however the libtinfo.so.5 is built agains the glibc-native version.  This
may introduce symbol dependencies that can not be satisfied by the host
system's glibc.  Using the LD_LIBRARY_PATH will cause the xsct component
and anything it calls to attempt to use this potentially incompatible
library.

 error loading hsi package: couldn't load file "libxv_commontasks.so":
       /lib/x86_64-linux-gnu/libc.so.6: version `GLIBC_2.33' not found
       (required by .../base-pdi/1.0-r0/recipe-sysroot-native/usr/lib/libtinfo.so.5)

Signed-off-by: Mark Hatle <mark.hatle@amd.com>