machine files: Additional version wrappers
In order to support multiple versions of XSCT, we need a way to trigger
slightly different versions of the machine.conf files to load.  A wrapper
that includes a versioned machine.conf will handle this along with warning
the user if the configuration is not valid.

These wrappers were missed in the initial set.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>