external-hdf: Update deployed output and other changes
do_install has been removed, there is no purpose to create a target package.
There is no target software that can directly consume an xsa file.

do_deploy has been updated to follow the standard deployment filename
practices including versioning the output file, and creating a series of links
to the latest version of the file.  The written file is always versioned, the
version-less access is always a symlink.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>