dtgen_dfx: Fix -processor and -processor_ip parameter
dtgen_dfx did not observe the -processor nor -processor_ip parameter but
instead used always the first processor returned by getprocessors.

This fails, if the first processor has not the correct/expected architecture.
For example, getprocessors could return "microblaze_I" as first element in a
design for a ZynqMP Ultrascale+ with an embedded Microblaze processor, even
though "-processor_ip psu_cortexa53" is being passed as parameter by the
recipe/Openembedded build system.

With this change the script uses:
(1) the -processor parameter to find the given instance directly
(2) the -processor_ip parameter to find instances with matching IP name
(3) fall back to the previous behaviour of selecting the first entry of
getprocessors.

Precedence is as listed, i.e. if -processor is given, -processor_ip is ignored.

This fixes #54

Signed-off-by: Hannes Weisbach <weisbach@neosat.de>